refactor: move CSV export logic from EmployeesTable to dedicated utility function for better separation of concerns
